update: 重构代码，更新依赖

+++ public/script.js
@@ -25,31 +25,8 @@ fileInput.addEventListener('change', () => {
         document.getElementById('tableContainer')?.style.setProperty('display', 'flex')
         let total = 0
         for (const file of fileInput.files) {
-            const newRow = fileList.insertRow()
-            const name = newRow.insertCell()
-            const size = newRow.insertCell()
-            const status = newRow.insertCell()
-            const action = newRow.insertCell()
-            name.textContent = file.name
-            size.textContent = filesize(file.size)
             total += file.size
-            if (file.size > 1e8)
-                status.textContent = statusMap.tooBig
-            else {
-                status.textContent = statusMap.notUploaded
-                name.contentEditable = true
-                const buttonElement = document.createElement('button')
-                buttonElement.textContent = actionMap.upload
-                buttonElement.classList.add('action')
-                action.appendChild(buttonElement)
-                buttonElement.onclick = () => {
-                    if (buttonElement.classList.contains('action')) {
-                        const tr = buttonElement.parentElement.parentElement
-                        const rowIndex = Array.from(tr.parentElement.children).indexOf(tr)
-                        toServer(rowIndex)
-                    }
-                }
-            }
+            insertRow(file)
         }
         document.getElementById('count').textContent = `Count: ${fileInput.files.length}`
         document.getElementById('total').textContent = `Total: ${filesize(total)}`
@@ -58,52 +35,62 @@ fileInput.addEventListener('change', () => {
     }
 })
 
-const toServer = (index) => {
-    const file = fileInput.files[index]
-    const [name, size, status, action] = fileList.rows[index].cells
-    switch (action.firstChild.textContent) {
-        case actionMap.upload:
-            action.firstChild.setAttribute("disabled", true)
-            status.textContent = statusMap.uploading
-            fetch(`/${name.textContent}`, {
-                method: 'PUT',
-                body: file
-            }).then(async (resp) => {
+const insertRow = (file) => {
+    const row = fileList.insertRow()
+    const name = row.insertCell()
+    const size = row.insertCell()
+    const status = row.insertCell()
+    const action = row.insertCell()
+    name.textContent = file.name
+    size.textContent = filesize(file.size)
+    if (file.size > 1e8)
+        status.textContent = statusMap.tooBig
+    else {
+        status.textContent = statusMap.notUploaded
+        name.contentEditable = true
+        const actionButton = document.createElement('button')
+        actionButton.textContent = actionMap.upload
+        actionButton.classList.add('action')
+        action.appendChild(actionButton)
+        actionButton.onclick = async () => {
+            if (actionButton.textContent === actionMap.upload) {
+                actionButton.setAttribute('disabled', true)
+                status.textContent = statusMap.uploading
+                const resp = await fetch(`/${file.name}`, {
+                    method: 'PUT',
+                    body: file
+                })
                 if (resp.status === 200) {
                     const data = await resp.json()
-                    const linkElement = document.createElement('a')
-                    linkElement.href = decodeURI(data.url)
-                    linkElement.textContent = name.textContent
-                    linkElement.onclick = (event) => {
-                        event.preventDefault()
+                    const nameLink = document.createElement('u')
+                    nameLink.textContent = file.name
+                    nameLink.onclick = () => {
                         navigator.clipboard.writeText(data.url)
                     }
                     name.textContent = ''
-                    name.appendChild(linkElement)
+                    name.appendChild(nameLink)
                     name.contentEditable = false
                     status.textContent = statusMap.uploaded
-                    action.firstChild.textContent = actionMap.delete
+                    actionButton.textContent = actionMap.delete
                 } else {
                     status.textContent = statusMap.uploadFail
                 }
-                action.firstChild.removeAttribute("disabled")
-            })
-            break
-        case actionMap.delete:
-            action.firstChild.setAttribute("disabled", true)
-            status.textContent = statusMap.deleting
-            fetch(`/${name.firstChild.textContent}`, {
-                method: 'DELETE'
-            }).then((resp) => {
+                actionButton.removeAttribute('disabled')
+            } else if (actionButton.textContent === actionMap.delete) {
+                actionButton.setAttribute('disabled', true)
+                status.textContent = statusMap.deleting
+                const resp = await fetch(`/${file.name}`, {
+                    method: 'DELETE'
+                })
                 if (resp.status === 200) {
-                    name.textContent = name.firstChild.textContent
+                    name.textContent = file.name
                     status.textContent = statusMap.deleted
-                    action.removeChild(action.firstChild)
+                    action.removeChild(actionButton)
                 } else {
                     status.textContent = statusMap.deleteFail
-                    action.firstChild.removeAttribute("disabled")
+                    actionButton.removeAttribute('disabled')
                 }
-            })
-            break
+            }
+        }
     }
 }
